nasa-backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
Pillow==10.0.1
redis==5.0.1
django-redis==5.3.0
requests==2.31.0
numpy==1.24.3
